requirements.txt
boto3==1.34.34
PyGithub==2.1.1
pyyaml==6.0.1
requests==2.31.0 